pyproject.toml
[project]
name = "dsrp-sentiment-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
    "python-dotenv>=1.2.1",
]
